🧹 Restructure CF functions code into smaller files
